requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.13.0
httpx==0.25.2
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
pg8000==1.31.2
